+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v0.2beta14 2026-01-07
+- Make chargee_charge_limit modifiable if it is a local config entity (local device entities are not modifiable).
 - Removed defaults for charger effective voltage and max current, ie. must be set before use.
 
 ## v0.2beta13 2026-01-06